+++ HISTORY.rst
@@ -7,3 +7,17 @@ History
 ---------------------
 
 * First release on Github.
+
+0.2.0 (2014-09-19)
+---------------------
+
+* Code cleanup and updated utils and traversals
+
+
+0.3.0 (2014-10-08)
+---------------------
+
+* Added the following traversals:
+    Pre-order
+    Post-order
+    Level-order